chore: Automate sandbox versioning and release creation workflow
Automate the sandbox versioning and release creation workflow to improve efficiency and reduce manual errors. This includes updating the workflow file "01-version-and-promotion.yml" to use the action "softprops/action-gh-release@v2.0.8" for creating GitHub releases. The release name is now "Sandbox Release" followed by the new version number. The release notes are generated automatically and the discussion category is set to "Releases". Additionally, the workflow now includes specific files to be included in the release, such as all files in the directory "${{ env.CONFIG_DIRECTORY }}/**/*.tf" and "README.md".